Pull from latest code Jan. 28, 2015

+++ autopack/Graphics.py
@@ -379,7 +379,9 @@ class AutopackViewer:
             if p is None :
                 p = self.vi.newEmpty(name,parent='O%s'%orga.name)
             print ("orga.representation ",name, p, orga.representation)
-            self.vi.reParent(orga.representation,p)  
+            op = self.vi.getObject(orga.representation)
+            if op is not None :
+                self.vi.reParent(orga.representation,p)  
     #        orga.ref_obj = name
 
     def createOrganelMesh(self, orga ):
@@ -762,10 +764,13 @@ class AutopackViewer:
                 #what about a reall long one -> cloner or instance
             else :
                 if self.vi.host.find("blender") != -1 :
+                    #is that not correctly scaled ?
                     circle = self.vi.build_2dshape(name+"_shape",opts=[ ingr.encapsulatingRadius,])[0]
                     extruder,shape = self.vi.extrudeSpline(snake,shape=circle,parent=parent)#shoud use the radius for a circle ?
                     #reparent ?
                     #should wereparent the extruder
+        #what about primitive display self.doSpheres
+
 
     def delIngredientGrow(self,ingr):
         print ("delIngrGrow",ingr,ingr.nbCurve)
@@ -816,7 +821,7 @@ class AutopackViewer:
                 
     def displayIngrediants(self,recipes):
         for ingr in recipes.ingredients:
-            if ingr.mesh == None :#mes_3d?
+            if type(ingr.mesh) == type(None) :#mes_3d?
                 #try get it
                 ingr.mesh = self.helper.getObject(ingr.name) 
             if ingr.mesh: # display mesh
@@ -928,6 +933,9 @@ class AutopackViewer:
                 radii.append( ingr.radii[level][ii] )
         mat = rot.copy()
         mat[:3, 3] = pos
+        if self.helper.instance_dupliFace :
+            mat = rot.copy().transpose()
+            mat[3][:3]=pos
         return verts,radii,numpy.array(mat)
 
     def replaceIngrMesh(self,ingredient, newMesh):
@@ -1004,11 +1012,16 @@ class AutopackViewer:
             if not hasattr(ingredient,'ipoly') or ingredient.ipoly is None or dejavui:
                 color = [ingredient.color] if ingredient.color is not None else None
 #                print o.name+self.histo.FillName[self.histo.cFill]+ingredient.name
-                ingredient.ipoly = self.vi.instancePolygon(o.name+self.histo.FillName[self.histo.cFill]+ingredient.name,
+                axis = numpy.array(ingredient.principalVector[:])
+#                if self.vi.host.find("blender") != -1 and self.vi.instance_dupliFace and ingredient.coordsystem == "left": 
+##                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
+#                    axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
+                print ("build ipoly ",ingredient.o_name)
+                ingredient.ipoly = self.vi.instancePolygon(o.name+self.histo.FillName[self.histo.cFill]+ingredient.o_name,
                                         matrices=matrices,
                                         mesh=polygon,parent = parent,
                                         transpose= True,colors=color,
-                                        axis=ingredient.principalVector)
+                                        axis=axis)
         
     
     def displayInstancesIngredient(self,ingredient, matrices):
@@ -1023,11 +1036,16 @@ class AutopackViewer:
             if parent is None :
                 parent=self.vi.newEmpty(name, parent=self.orgaToMasterGeom[ingredient])
 #                    self.vi.AddObject(parent)
-            ingredient.ipoly = self.vi.instancePolygon(self.histo.FillName[self.histo.cFill]+ingredient.name,
+            if not hasattr(ingredient,'ipoly') or ingredient.ipoly is None :
+                axis = numpy.array(ingredient.principalVector[:])
+#                if self.vi.host.find("blender") != -1 and self.vi.instance_dupliFace and ingredient.coordsystem == "left": 
+    #                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
+#                    axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
+                ingredient.ipoly = self.vi.instancePolygon(self.histo.FillName[self.histo.cFill]+ingredient.name,
                                         matrices=matrices,
                                         mesh=polygon,parent = parent,
                                         transpose= True,colors=[ingredient.color],
-                                        axis=ingredient.principalVector)
+                                        axis=axis)
         
     def displayCytoplasmIngredients(self):
         verts = {}
@@ -1075,14 +1093,16 @@ class AutopackViewer:
             meshGeoms = {}#self.meshGeoms#{}
             inds = {}
             for pos, rot, ingr, ptInd in self.histo.molecules:
-                #print "ingr",ingr,ingr.mesh,ingr.mesh_3d                
+#                print ("ingr",ingr,ingr.mesh,ingr.mesh_3d)               
                 if ingr.mesh: # display mesh
-#                    geom = ingr.mesh
                     mat = rot.copy()
                     mat[:3, 3] = pos
-                    if self.helper.host == 'dejavu':
-                        mry90 = self.helper.rotation_matrix(-math.pi/2.0, [0.0,1.0,0.0])#?
-                        mat = numpy.array(numpy.matrix(mat)*numpy.matrix(mry90))
+#                    if self.helper.host == 'dejavu':
+#                        mry90 = self.helper.rotation_matrix(-math.pi/2.0, [0.0,1.0,0.0])#?
+#                        mat = numpy.array(numpy.matrix(mat)*numpy.matrix(mry90))
+                    if self.helper.instance_dupliFace :
+                        mat = rot.copy().transpose()
+                        mat[3][:3]=pos
                     if ingr not in meshGeoms:
                         inds[ingr] = [ptInd]
                         meshGeoms[ingr] = [mat]
@@ -1097,51 +1117,42 @@ class AutopackViewer:
                         #    self.vi.AddObject(geom, parent=self.orgaToMasterGeom[ingr])
             j=0
             for ingr in r.ingredients:
-                #self.vi.progressBar(j/ningr,label="instances for "+str(j)+" "+ingr.name+" "+str(len(meshGeoms[ingr]))) 
+#                print("process ",ingr.o_name) 
                 geom = ingr.mesh
-                #geom.Set(instanceMatrices=meshGeoms[ingr], visible=1)        
-                #if self.ViewerType != 'dejavu':
-                    #find the polygon and the ingr?#polygon = ingr.mesh_3d
                 polygon = ingr.mesh_3d
                 name = "Meshs_"+ingr.name.replace(" ","_")
                 parent = self.vi.getObject(name)
                 if parent is None :
                     parent=self.vi.newEmpty(name, parent=self.orgaToMasterGeom[ingr])
-#                    self.vi.AddObject(parent)
                 if self.helper.host == 'dejavu':
                     parent = None
                 if ingr not in meshGeoms:
+                    print ("ingr not in meshGeoms",ingr,meshGeoms)
                     continue
                 axis = numpy.array(ingr.principalVector[:])
-                if self.vi.host.find("blender") != -1 and self.vi.dupliVert : 
-                    if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
-                        axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
-                        print (self.helper.getType(self.helper.getChilds(polygon)[0]))
-                ingr.ipoly = self.vi.instancePolygon("cyto_"+self.histo.FillName[self.histo.cFill]+ingr.name,
+#                if self.vi.host.find("blender") != -1 and self.vi.instance_dupliFace and ingr.coordsystem == "left": 
+#                    axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
+#                print ("make ipoly",)
+                ingr.ipoly = self.vi.instancePolygon("cyto_"+self.histo.FillName[self.histo.cFill]+ingr.o_name,
                                             matrices=meshGeoms[ingr],
                                             mesh=polygon,parent = parent,
                                             transpose= True,colors=[ingr.color],
                                             axis=axis)
-#                if self.doOrder :
-                #print ingr.ipoly
                 if self.vi.host.find("blender") != -1 :
                     self.orgaToMasterGeom[ingr] = polygon
-                    if not self.vi.dupliVert :
+                    if not self.vi.instance_dupliFace :
                         self.vi.setLayers(polygon,[1])#and do it for child too.
-#                    else :    
-#                        if ingr.coordsystem == "left":
-#                            self.vi.rotateObj(polygon,[0.0,-math.pi/2.0,0.0])
-
                 elif self.vi.host == "dejavu" or self.vi.host == "softimage":
                     self.orgaToMasterGeom[ingr] = ingr.mesh
                 if self.helper.host != 'dejavu':
-                    if type(ingr.ipoly) != list : return
-                    for i,ip in enumerate(ingr.ipoly):
-    #                    print i,ip,type(ip)
-                        name = self.vi.getName(ip)
-                        if inds[ingr][i] in self.histo.order:
-    #                        print name,name+"_"+str(self.histo.order[inds[ingr][i]])
-                            self.vi.setName(ip,name+"_"+str(self.histo.order[inds[ingr][i]]))
+                    if not self.helper.instance_dupliFace:
+                        if type(ingr.ipoly) != list  : return
+                        for i,ip in enumerate(ingr.ipoly):
+        #                    print i,ip,type(ip)
+                            name = self.vi.getName(ip)
+                            if inds[ingr][i] in self.histo.order:
+        #                        print name,name+"_"+str(self.histo.order[inds[ingr][i]])
+                                self.vi.setName(ip,name+"_"+str(self.histo.order[inds[ingr][i]]))
 #                self.vi.progressBar(j/ningr,label="instances for "+str(j)+" "+ingr.name+" "+str(len(meshGeoms[ingr]))) 
                 j+=1
                 
@@ -1213,10 +1224,13 @@ class AutopackViewer:
                     #print pos                          
                     mat = rot.copy()
                     mat[:3, 3] = pos
-                    if self.helper.host == 'dejavu':
-                        mry90 = self.helper.rotation_matrix(-math.pi/2.0, [0.0,1.0,0.0])
-                        mat = numpy.array(numpy.matrix(mat)*numpy.matrix(mry90))
-
+#                    if self.helper.host == 'dejavu':
+#                        mry90 = self.helper.rotation_matrix(-math.pi/2.0, [0.0,1.0,0.0])
+#                        mat = numpy.array(numpy.matrix(mat)*numpy.matrix(mry90))
+                    if self.helper.instance_dupliFace :
+#                        self.helper.host.find("blender") != -1 :
+                        mat = rot.copy().transpose()
+                        mat[3][:3]=pos
                     if ingr not in matrices:
                         matrices[ingr] = []
                     matrices[ingr].append(mat)
@@ -1253,20 +1267,21 @@ class AutopackViewer:
                             parent = None                           
                         print("ri instanciation of polygon",polygon)
                         axis = numpy.array(ingr.principalVector[:])
-                        if self.vi.host.find("blender") != -1 and self.vi.dupliVert : 
-                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
-                                axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
-                                print (self.helper.getType(self.helper.getChilds(polygon)[0]))
+#                        if self.vi.host.find("blender") != -1 and self.vi.instance_dupliFace and ingr.coordsystem == "left": 
+#                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
+#                            axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
+#                            print (self.helper.getType(self.helper.getChilds(polygon)[0]))
                         ingr.ipoly = self.vi.instancePolygon(orga.name+self.histo.FillName[self.histo.cFill]+ingr.name,
                                                     matrices=matrices[ingr],
                                                     mesh=polygon,
                                                     parent = parent,
                                                     transpose=True,colors=[ingr.color],
                                                     axis=axis)
+                        #print ("ingr instance build, reparent".ingr.ipoly,self.orgaToMasterGeom[ingr])
                         #principal vector rotate by 90degree for blender dupliVert?
                         if self.vi.host.find("blender") != -1 :
                             self.orgaToMasterGeom[ingr] = polygon
-                            if not self.vi.dupliVert : 
+                            if not self.vi.instance_dupliFace : 
                                 self.vi.setLayers(polygon,[1])#and do it for child too.
 #                            else :    
 #                                if ingr.coordsystem == "left":
@@ -1302,19 +1317,20 @@ class AutopackViewer:
                             parent = None                           
                         print("rs instanciation of polygon",polygon)
                         axis = numpy.array(ingr.principalVector[:])
-                        if self.vi.host.find("blender") != -1 and self.vi.dupliVert : 
-                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
-                                axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
-                                print (self.helper.getType(self.helper.getChilds(polygon)[0]))
+#                        if self.vi.host.find("blender") != -1 and self.vi.instance_dupliFace and ingr.coordsystem == "left": 
+#                            if self.helper.getType(self.helper.getChilds(polygon)[0]) != self.helper.EMPTY:
+#                            axis = self.vi.rotatePoint(axis,[0.,0.,0.],[0.0,1.0,0.0,-math.pi/2.0])
+#                                print (self.helper.getType(self.helper.getChilds(polygon)[0]))
                         ingr.ipoly = self.vi.instancePolygon(orga.name+self.histo.FillName[self.histo.cFill]+ingr.name,
                                                     matrices=matrices[ingr],
                                                     mesh=polygon,
                                                     parent = parent,
                                                     transpose= True,colors=[ingr.color],
-                                                    axis=axis)            
+                                                    axis=axis) 
+                        #print ("ingr instance build, reparent".ingr.ipoly,self.orgaToMasterGeom[ingr])
                         if self.vi.host.find("blender") != -1 :
                             self.orgaToMasterGeom[ingr] = polygon
-                            if not self.vi.dupliVert :
+                            if not self.vi.instance_dupliFace :
                                 self.vi.setLayers(polygon,[1])#and do it for child too.
 #                            else :    
 #                                if ingr.coordsystem == "left":
@@ -2182,11 +2198,13 @@ class AutopackViewer:
                 
     def clearRecipe(self,recipe,*args):
         """ will clear everything related to self.recipe"""
-        parent = self.helper.getObject(recipe)
-        if parent is not None :
-            instances = self.helper.getChilds(parent)
-            [self.helper.deleteObject(o) for o in instances]
-            self.helper.deleteObject(parent)
+        if not self.helper.instance_dupliFace :
+            parent = self.helper.getObject(recipe)
+            if parent is not None :
+                instances = self.helper.getChilds(parent)
+                [self.helper.deleteObject(o) for o in instances]
+                self.helper.deleteObject(parent)
+        
 
     def clearAll(self,recipe):
         #shoud remove everything
